Create test suite for metrics publisher tests
